Interpreter runs as fast as the clock frequency given in the constructor

+++ src/cores/SCC68070/SCC68070.hpp
@@ -227,7 +227,7 @@ public:
     void RebootCore();
     void SingleStep();
     void ResetOperation();
-    unsigned long long instructionCount;
+    uint64_t instructionCount;
 
 private:
     VDSC* vdsc;
